+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.6.0] - 2022-03-26
+
 ### Added
 
 - Add `apilytics.flask.apilytics_middleware` for sending metrics from Flask applications.